An autoclicker made to feel like a minigame
This is a beta. The code can be simplified, and the mods lacks features (like a in-game option menu. Or even better, a minigame tab like every other minigame). But the concept is here, and I think I like the progress curve I chosed. I like it, and I don't feel the need to push that last 20% of work, but if you guys like it I will.
Any feedback is welcomed